Create Python File: Check the Observability
In the code below, we will use the “obsv()” function from the python controls library. Here, we check the rank of the observability matrix returned by the function. We then also check its determinant for non-singularity. The rank of the observability matrix is full. So the system is observable. The determinant of the matrix is: 24.06
We can see that the system is observable as well. The observability matrix O is also full rank. Further, the determinant of matrix O is non zero. So with the above parameters, the system is controllable and observable.